Move `toplevelprop` to `spec/toplevel`
